capacidad cambio de tipo de dato a Long

+++ src/main/java/daw2a/gestion_alimentos_api_rest/dto/ubicacion/CrearUbicacionDTO.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class CrearUbicacionDTO {
     private String descripcion;
     private String tipoUbicacion;
-    private int capacidad;
+    private Long capacidad;
 }
